+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed overallMaxLevel setting
+- Fixed multiplayer enchantment sync
 ### Changed:
 -